move PipelineWorker from cluster to automation
